requirements.txt
pandas
python-dotenv
tqdm
httpx
pandas-stubs
types-tqdm
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-api-python-client-stubs
google-auth-stubs